chore: move swagger to lib
